Persist console messages across reconnects
