+++ lib/cli.js
@@ -44,6 +44,7 @@ const resolvedArgv = require.resolve(path.resolve(process.cwd(), config._.shift(
 process.argv = [ process.argv[0], resolvedArgv ].concat(config._);
 
 const agents = startDebug(config);
+agents.get('Console'); // To make sure we capture console messages
 agents.get('Debugger'); // To make sure we capture break events
 agents.get('Worker'); // To make sure we capture workers that are created
